+++ frontend/src/pages/GroupInitPage/components/GroupInitHeader/GroupInitHeader.styles.tsx
